Fixed a bug in liftPar, added liftParErase.
* Fixed a bug which caused errouneous results in liftPar if the pipe's
outputs weren't unique.
* Added history-erasing versions of liftPar and liftPar': liftParErase
and liftParErase'. These are useful when one needs to e.g. sort the
output: when filename F_i becomes F_j through sorting, we really wish to
work with F_j from that point on. The fact that F_i was at position i in
the index prior to sorting is no longer relevant.
* Added the option to sort files descendingly the pipe 'sortByDate' in
Main.hs